weDevs settings api fixes
