+++ composer.json
@@ -11,7 +11,7 @@
   ],
   "minimum-stability": "dev",
   "require": {
-    "php": ">=5.4.0",
+    "php": ">=5.5",
     "composer/installers": ">=1.4",
     "appsero/client": "dev-develop"
   },
